requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==23.2.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
arm_pytorch_utilities==0.4.3
asttokens==3.0.0
attrs==25.2.0
beautifulsoup4==4.13.4
blinker==1.9.0
casadi==3.6.7
certifi==2025.1.31
charset-normalizer==3.4.1
chumpy==0.70
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
cyclonedds==0.10.2
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.2.1
easydict==1.13
etils==1.12.2
evdev==1.9.1
exceptiongroup==1.2.2
executing==2.2.0
Farama-Notifications==0.0.4
fastapi==0.115.12
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.18.0
Flask==3.0.3
fonttools==4.56.0
fsspec==2025.3.0
gdown==5.2.0
glfw==2.8.0
gradio==5.23.1
gradio_client==1.8.0
groovy==0.1.2
gymnasium==1.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipdb==0.13.13
ipython==8.34.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
loguru==0.7.3
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.5.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.2.2
mpmath==1.3.0
mujoco==3.3.0
narwhals==1.30.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
numpy-stl==3.2.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
open3d==0.19.0
opencv-python==4.11.0.86
orjson==3.10.16
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pybullet==3.2.7
pydantic==2.11.1
pydantic_core==2.33.0
pydub==0.25.1
Pygments==2.19.1
pynput==1.8.1
PyOpenGL==3.1.9
pyparsing==3.2.1
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-utils==3.9.1
python-xlib==0.33
pytorch-kinematics==0.7.5
pytorch-seed==0.2.0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rerun-sdk==0.22.0
retrying==1.3.4
rich==13.9.4
rich-click==1.8.8
rpds-py==0.23.1
ruff==0.11.2
safehttpx==0.1.6
scikit-learn==1.6.1
# scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy-split_1739790642651/work/dist/scipy-1.15.2-cp310-cp310-linux_x86_64.whl#sha256=9e52bad6c3294d1a5b04a13632118ca2157130603c6c018c2d710162b223b27e
semantic-version==2.10.0
setuptools-scm==8.2.0
shellingham==1.5.4
six==1.17.0
# smpl_sim @ git+https://github.com/ZhengyiLuo/SMPLSim.git@c9e7be186f2b395950cdaab19db4b07fd7ae895c
smplx @ git+https://github.com/ZhengyiLuo/smplx.git@a5b8e4ac14f79f3f33fd2cf2a16e6f507146b813
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.1
threadpoolctl==3.6.0
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0
torchgeometry==0.1.2
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.6.4
triton==3.2.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
-e git+https://github.com/unitreerobotics/unitree_sdk2_python.git@986f39d54182badc1aa3a0c282bcd898fba4ef20#egg=unitree_sdk2py
urllib3==2.3.0
uvicorn==0.34.0
vtk==9.4.1
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.0.6
widgetsnbextension==4.0.13
zipp==3.21.0
